+++ apollo-server-app/src/typeDefs/ApiSrmSchema.js
@@ -119,6 +119,25 @@ const apiSrmType = gql`
     user: Int,
     confirmed_payments: [Int]
   }
+  type Bank {
+    id: ID,
+    created: String,
+    updated: String,
+    name: String,
+    name_institution: String,
+    bank_code: String,
+    active: Boolean,
+    bank_alias: Int
+  }
+  type BankAccountType {
+    id: ID,
+    created: String,
+    updated: String,
+    type: String
+  }
+  type AttemptResponse {
+    message: String
+  }
   input EnterpriseInput {
     name: String!
     business_name: String!
@@ -202,8 +221,20 @@ const apiSrmType = gql`
     confirmed_payments: [Int]!
     bank_account: Int!
   }
+  input AttemptStateInput {
+    attempt_id: Int!
+    state: String!
+  }
+  input BankAccountInput {
+    provider_id: ID!
+    account_number: String!
+    email: String!
+    account_holder_name: String!
+    account_holder_document_number: String!
+    bank: Int!
+    account_type: Int!
+  }
   type Query {
-    Services: Available
     representatives: [Representative]
     payerCompanyForRepresentative(
       input: RepresentantiveInput
@@ -219,10 +250,13 @@ const apiSrmType = gql`
     getPayerListFromProvider(input: payerProviderInput): [Enterprise]
     getProviderListFromPayer(input: payerProviderInput): [Enterprise]
     advanceSimulation(input: AdvanceSimulationInput): AdvanceSimulation
+    bank:[Bank]
+    bankAccountType:[BankAccountType]
   }
   extend type Mutation {
     enterprises(input: EnterpriseInput): Enterprise
     newEnterpriseProvider(enterprise_id: Int!): enterpriseAssociation
+    newProviderBankAccount(input: BankAccountInput): [BankAccount]
     newEnterprisePayer(enterprise_id: Int!): enterprisePayerAssociation
     asssignRepresentativeToPayerEnterprise(
       input: AssignPayerRepresentativeInput
@@ -239,6 +273,10 @@ const apiSrmType = gql`
       input: setRepresentativeFavoriteProviderInput
     ): Representative
     confirmedPaymentAdvanceAttempt(input: AdvanceAttemptInput): AdvanceAttempt
+    confirmedPaymentAttemptState(input: AttemptStateInput  ): AttemptResponse
+    
+
+    
   }
 `;